requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
drf-yasg==1.21.11
python-decouple==3.8
requests==2.32.5
dj-database-url==2.1.0
gunicorn==21.2.0
